requirements.txt
streamlit==1.37.1
neo4j==5.24.0
python-dotenv==1.0.1
bcrypt==4.2.0
pandas==2.2.2
